+++ Dto/DtoResolverInterface.php
@@ -18,11 +18,18 @@ use JsonSerializable;
 interface DtoResolverInterface extends JsonSerializable
 {
     /**
-     * Returns DTO properties as array
+     * Returns Dto properties as array
      *
      * @param bool $onlyDefinedData
      *
      * @return array
      */
     public function toArray(bool $onlyDefinedData = true): array;
+
+    /**
+     * Returns Dto initial properties as array
+     *
+     * @return array
+     */
+    public function getDefinedProperties(): array;
 }